discord_bot/requirements.txt
discord.py==2.4.0
python-dotenv==1.0.1
requests==2.32.3
uvicorn==0.34.0
fastapi==0.115.8